Jam changes for building on the AMD64 platform.  The .NET 2005 beta2 compiler requires several flags.  First it is necessary to turn off the depreciation stuff, CRT functions like strcpy sprintf and so on are now viewed as insecure by MS.  For now it is still possible to use them with the correct application of compile flags.  MS is making noises in line with taking these functions away for good.
The "Win64" token was used for 64 bit builds.  This is now a
bit ambiguous since it covers x64, ia64 and em64t.  With this
change Win64 has been retired and we use x64 for amd64, ia64
for the Itanium and em64t for the extended memory processors.
Set MSVCVer accordingly.  MSVCNT is set as always, to the
location of the VC directory, VC98, VC7 and now just VC.

For the most part the x64 build works on amd64, the exception
being a template problem in nettcp.cc.  That problem will not
be corrected in this change.


p4transfer.py: Transferred from production

[git-p4: depot-paths = "//guest/perforce_software/jam/": change = 9938]